% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixVizR-package.R
\docType{package}
\name{omixVizR-package}
\alias{omixVizR}
\alias{omixVizR-package}
\title{omixVizR: Omics Visualization Toolkit}
\description{
The omixVizR package provides tools for visualizing omics data, including genomic, transcriptomic, and proteomic data. The package is designed to facilitate the exploration and interpretation of complex omics datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leslie-lu.github.io/}
}

}
\author{
\strong{Maintainer}: Zhen Lu \email{luzh29@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0002-3481-6310}{ORCID})

}
\keyword{internal}
