% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixVizR-package.R
\docType{package}
\name{omixVizR-package}
\alias{omixVizR}
\alias{omixVizR-package}
\title{omixVizR: A Toolkit for Omics Data Visualization}
\description{
Provides a suite of tools for the comprehensive visualization of multi-omics data, including genomics, transcriptomics, and proteomics. Offers user-friendly functions to generate publication-quality plots, thereby facilitating the exploration and interpretation of complex biological datasets. Supports seamless integration with popular R visualization frameworks and is well-suited for both exploratory data analysis and the presentation of final results. Key formats and methods are presented in Huang, S., et al. (2024) "The Born in Guangzhou Cohort Study enables generational genetic discoveries" \doi{10.1038/s41586-023-06988-4}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Leslie-Lu/omixVizR}
  \item Report bugs at \url{https://github.com/Leslie-Lu/omixVizR/issues}
}

}
\author{
\strong{Maintainer}: Zhen Lu \email{luzh29@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0002-3481-6310}{ORCID})

}
\keyword{internal}
