% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omixVizR-package.R
\docType{package}
\name{omixVizR-package}
\alias{omixVizR}
\alias{omixVizR-package}
\title{omixVizR: 'omixVizR': Omics Visualization Toolkit}
\description{
'omixVizR' is an R package for comprehensive visualization of multi-omics data, including genomics, transcriptomics, and proteomics. It provides user-friendly functions to generate publication-quality plots, facilitating the exploration and interpretation of complex biological datasets. The package supports integration with popular R visualization frameworks and is suitable for both exploratory data analysis and result presentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://leslie-lu.github.io/}
}

}
\author{
\strong{Maintainer}: Zhen Lu \email{luzh29@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0002-3481-6310}{ORCID})

}
\keyword{internal}
